requirements.txt
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
APScheduler==3.10.4
Async-OAuthlib==0.0.9
attrs==23.1.0
black==23.11.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord==2.3.2
discord.py==2.3.2
distlib==0.3.7
filelock==3.13.1
flake8==6.1.0
Flask==3.0.0
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
priority==2.0.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pyproject-api==1.6.1
python-dotenv==1.0.0
pytz==2023.3.post1
Quart==0.19.4

requests==2.31.0
six==1.16.0
tox==4.11.3
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
virtualenv==20.24.6
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.2
